Support for :skipif: in testoutput directive
The value of the skipif option is evaluated as a Python expression. If
the result is a true-ish value, the directive is omitted from the test
run just as if it wasn’t present in the file at all.

As said in the subject-line, currently this is only implemented for the
testoutput directive.

Closes: #17