WIP Support for :skipif:
* currently only supported in the testoutput directive
* completely skips the doctest if the expression in :skipif: evaluates
  to True.

It is planned to extend this commit s.t. the behavior of skipif is the
same as writtin the documentation of sphinx.ext.doctest:

The value of the skipif option is evaluated as a Python expression. If
the result is a true value, the directive is omitted from the test run
just as if it wasn’t present in the file at all.