Adapt to option parsing behavior of sphinx
sphinx-build -M doctest requires that there is a newline between an
option block and a text block (code/output). Otherwise an "invalid
option block" warning is output. Furthermore, spinx-build -M doctest
also outputs a warning if the text block is empty.

pytest-sphinx raises a ValueError if the text inside the
directives is not properly formatted.